requirements.txt
opencv-python-headless>=4.5.0
numpy>=1.21.0
torch>=1.9.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.10.0 --index-url https://download.pytorch.org/whl/cpu
matplotlib>=3.3.0
scipy>=1.7.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
trimesh>=3.9.0
Pillow>=8.0.0
transformers>=4.11.0
flask==2.3.2
flask-cors==4.0.0